YouseAI/requirements.txt
asgiref==3.7.2
cffi==1.15.1
channels==3.0.4 
cryptography==40.0.2
Django==4.2.1

django-rest-knox==4.2.0
djangorestframework==3.14.0
Pillow==9.5.0
pycparser==2.21
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.2
tzdata==2023.3

